move CountOf and Having specifications
