+++ src/QueryModifier/CountOf.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Happyr\DoctrineSpecification\Specification;
+namespace Happyr\DoctrineSpecification\QueryModifier;
 
 use Doctrine\ORM\QueryBuilder;
+use Happyr\DoctrineSpecification\Specification;
 
 /**
  * @author Tobias Nyholm
